Check if device is an iPad on safari
Unfortunately the latest iPad safari browser uses the same machintosh user agent. 
So to verify correctly if it's an iPad we add an extra check on the touchscreen; the iPad is the only one apple device wich use the mac User Agent and wich also has the touch screen events enabled.